pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "pwn-solver"
version = "0.1.0"
description = "Agentic LLM system for binary exploitation, powered by MCP tool servers and RAG"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "anthropic>=0.42.0",
    "mcp>=1.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
pwn = [
    "pwntools>=4.12.0",
    "ropper>=1.13.0",
]
rag = [
    "qdrant-client>=1.9.0",
    "voyageai>=0.3.0",
    "cohere>=5.0.0",
    "beautifulsoup4>=4.12.0",
    "html2text>=2024.2.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[project.scripts]
pwn-solver = "agent.cli:main"

[tool.setuptools.packages.find]
include = ["agent*", "mcp_servers*", "knowledge*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
